requirements.txt
transformers==4.46.2
torch
python-dotenv==1.0.1
accelerate>=0.26.0
Pillow
ago==0.0.95
async-timeout==5.0.1
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
behave==1.2.6
boto3==1.35.68
botocore==1.35.68
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
defusedxml==0.7.1
dotmap==1.3.30
elastic-transport==8.15.1
elasticsearch==8.16.0
exceptiongroup==1.2.2
execnet==2.1.1
fasteners==0.19
faust-cchardet==2.1.19
feedparser==6.0.11
filelock==3.16.1
fsspec==2024.10.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-vision==3.8.1
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
hjson==3.1.0
huggingface-hub==0.26.2
hurry.filesize==0.9
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
langdetect==1.0.9
lxml==5.3.0
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
mycdp==1.1.0
networkx==3.4.2
-e git+https://github.com/phanhoang1803/DKH-Thesis.git@e26cdcf2c7477ff9a43e3e2de07e354cd8a10613#egg=news_please&subdirectory=news-please
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
parsel==1.9.1
pdbp==1.6.1
pillow==11.0.0
plac==1.4.3
platformdirs==4.3.6
pluggy==1.5.0
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.18.0
PyMySQL==1.1.1
pynose==1.5.3
pyOpenSSL==24.2.1
pyotp==2.9.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.3
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
queuelib==1.7.0
readability-lxml==0.8.1
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rsa==4.9
s3transfer==0.10.4
safetensors==0.4.5
sbvirtualdisplay==1.3.0
Scrapy==2.12.0
selenium==4.26.1
seleniumbase==4.32.12
service-identity==24.2.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sympy==1.13.1
tabcompleter==1.4.0
tldextract==5.1.3
tokenizers==0.20.3
tomli==2.1.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.10.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
w3lib==2.2.1
warcio==1.7.4
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
zope.interface==7.1.1
transformers==4.46.2
torch
python-dotenv==1.0.1
accelerate>=0.26.0